+++ src/main/java/org/plumelib/util/CollectionsPlume.java
@@ -1,5 +1,4 @@
 // If you edit this file, you must also edit its tests.
-// For tests of this and the entire plume package, see class TestPlume.
 
 package org.plumelib.util;