+++ src/test/java/org/plumelib/SystemPlumeTest.java
@@ -1,0 +1,10 @@
+package org.plumelib.util;
+
+@SuppressWarnings({
+  "UseCorrectAssertInTests" // `assert` works fine in tests
+})
+public final class SystemPlumeTest {
+
+  // No tests. :-(
+
+}